編輯bug修正

+++ app/Http/Controllers/TurnoutsController.php
@@ -91,7 +91,8 @@ class TurnoutsController extends Controller
         //把要更新的檔案及舊檔名放入陣列。
         foreach (range(1,10) as $key) {
             if($request->hasFile('fileName'.$key)){
-                $old_file_arr   = array_add($old_file_arr , $key   , $item['fileName'.$key]);
+                if(!empty($item['fileName'.$key]))
+                    $old_file_arr   = array_add($old_file_arr , $key   , $item['fileName'.$key]);
                 $new_file_arr   = array_add($new_file_arr , $key   , $request->file('fileName'.$key)); //取新檔案
             }
         }
@@ -102,7 +103,8 @@ class TurnoutsController extends Controller
         foreach($new_file_arr as $key => $file){
             $original_name = $file->getClientOriginalName();  
             if($file->isValid()){                                                           //若新資料為有效資料
-                Storage::delete('Filebase/'.$old_file_arr[$key]);                           //再把舊檔刪除                 
+                if(Storage::disk('local')->has('/Filebase/'.$old_file_arr[$key]))           //舊檔案是否存在
+                    Storage::delete('Filebase/'.$old_file_arr[$key]);                       //再把舊檔刪除                 
                 $file->move(storage_path('app/Filebase/'),$original_name);                  //然後新檔移入
                 $item->update(['fileName'.$key => $original_name]);                         //將暫存檔名更換為真檔名
             }
